Rename 'grafana_server' to 'grafana_client'
